evc_hybrid_chat_v3/requirements.txt

pyyaml>=6.0.1
streamlit>=1.36.0
matplotlib>=3.8.0
